extract editor layers into seperate headers
