+++ src/game/editor/mapitems/layer_sounds.cpp
@@ -1,3 +1,5 @@
+#include "layer_sounds.h"
+
 #include <game/editor/editor.h>
 
 #include <game/generated/client_data.h>
