+++ src/game/editor/mapitems/layer_tiles.cpp
@@ -1,5 +1,7 @@
 /* (c) Magnus Auvinen. See licence.txt in the root of the distribution for more information. */
 /* If you are missing that file, acquire a complete release at teeworlds.com.                */
+#include "layer_tiles.h"
+
 #include <game/editor/editor.h>
 
 #include <engine/keys.h>
